vendor: Bump controller-tools to latest
This bump brings controller-tools up-to-date with the upstream master
branch as of Sept 6, 2020, and adds support for the following:

 - Ability to request a top-level "description" field to the validation
   schema. See further below for the rationale of why this was added.
 - Ability to request a top-level "x-preserve-unknown-fields" field to
   the validation schema. This allows CNPs and CCNPs to prevent their
   `spec` or `specs` from being a null object, which K8s will prune.
   This allows us to the maintain behavior with the CNP & CCNPS under
   v1beta1.
 - Ability to request anyOf validation. This currently doesn't have a
   use case, but was trivial enough to add.

To install a CNP or CCNP that has a top-level description field (see
example below), the "description" field must be present in the
validation schema. This field in the validation schema is now required
because v1 CRDs disallow "unknown" fields. The presence of a field in
the validation schema makes the apiserver aware of this field so that it
cannot be deemed "unknown". This was not required previously, as v1beta1
CRDs permitted "unknown" fields. In upcoming commits, we will move to v1
CRDs, which disallows "unknown" fields (`preserveUnknownFields`), hence
why the support was added and updated via this commit.

Note that this is actually non-standard, as we already have a free-form
field also called "description" under the CNP / CCNP `spec` for this
purpose, which should be preferred. See
https://github.com/cilium/cilium/issues/13155 for more information.

Example CNP / CCNP:

```
apiVersion: "cilium.io/v2"
kind: CiliumNetworkPolicy
description: "L7 policy for getting started using Kubernetes guide"
metadata:
  name: "l7-policy"
spec:
  endpointSelector:
    matchLabels:
      id: app1
  ingress:
  - fromEndpoints:
    - matchLabels:
        id: app2
    toPorts:
    - ports:
      - port: "80"
        protocol: TCP
      rules:
        http:
        - method: "GET"
          path: "/public"
```

Signed-off-by: Chris Tarazi <chris@isovalent.com>